requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
polars==0.20.3
pyarrow==14.0.2

# Database Connections
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
sqlalchemy==2.0.25

# Apache Spark for Big Data
pyspark==3.5.0

# Data Quality & Validation
great-expectations==0.18.8
pandera==0.18.0
pydantic==2.5.3

# API & Data Collection
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0
beautifulsoup4==4.12.2

# Streaming & Message Queue
kafka-python==2.0.2
confluent-kafka==2.3.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2
imbalanced-learn==0.11.0

# Deep Learning (Optional)
# tensorflow==2.15.0
# torch==2.1.2
# transformers==4.36.2

# Model Interpretability
shap==0.44.0
lime==0.2.0.1
alibi==0.9.4

# Statistical Analysis
scipy==1.11.4
statsmodels==0.14.1
lifelines==0.28.0  # Survival analysis

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
dash==2.14.2
streamlit==1.29.0
folium==0.15.1
geopandas==0.14.2

# Workflow Orchestration
apache-airflow==2.8.0
prefect==2.14.11

# Data Transformation
dbt-core==1.7.4
dbt-postgres==1.7.4

# Time Series Analysis
prophet==1.1.5
pmdarima==2.0.4

# Natural Language Processing (for clinical notes)
nltk==3.8.1
spacy==3.7.2

# API Development
fastapi==0.109.0
uvicorn==0.25.0
pydantic-settings==2.1.0

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.3
faker==22.0.0  # Generate synthetic data

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Jupyter & Notebooks
jupyter==1.0.0
jupyterlab==4.0.10
ipykernel==6.28.0
ipywidgets==8.1.1

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
joblib==1.3.2
click==8.1.7

# Geospatial Analysis
shapely==2.0.2
pyproj==3.6.1

# Excel & Report Generation
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.8
jinja2==3.1.2

# Cloud Storage (Optional)
# boto3==1.34.18  # AWS S3
# google-cloud-storage==2.14.0  # Google Cloud Storage
# azure-storage-blob==12.19.0  # Azure Blob Storage

# Performance
numba==0.58.1
dask==2024.1.0

# Data Encryption & Security
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

